Tests/LibWeb/Text/expected/abortsignal-timeout.txt

Time passed before abort event fired is at least 10 milliseconds: true
Reason type: TimeoutError
onabort event isTrusted: true
